doradjeno i skoro sve testirano

+++ src/main/java/com/example/project/controllers/CategoryController.java
@@ -13,45 +13,54 @@ import com.example.project.entities.CategoryEntity;
 import com.example.project.repositories.CategoryRepository;
 
 @RestController
-@RequestMapping(value = "/project/categories")
+@RequestMapping(value = "/api/v1/project/categories")
 public class CategoryController {
 
 	@Autowired
 	private CategoryRepository categoryRepository;
 
-	// Vrati sve kategorije
+	// Vrati sve kategorije TESTIRAO
 	@RequestMapping(value = "/", method = RequestMethod.GET)
 	public List<CategoryEntity> getCategories() {
 		return (List<CategoryEntity>) categoryRepository.findAll();
 	}
 
-	// Dodaj novu kategoriju
+	// Dodaj novu kategoriju TESTIRAO
 	@RequestMapping(value = "/", method = RequestMethod.POST)
 	public CategoryEntity addCategory(@RequestBody CategoryEntity category) {
 		return categoryRepository.save(category);
 	}
 
-	// Azuriraj kategoriju po ID-u
+	// Azuriraj kategoriju po ID-u TESTIRAO
 	@RequestMapping(value = "/{id}", method = RequestMethod.PUT)
 	public CategoryEntity updateCategory(@PathVariable Integer id, @RequestBody CategoryEntity cat) {
-		CategoryEntity category = categoryRepository.findById(id).get();
-		category.setCategoryName(cat.getCategoryName());
-		category.setCategoryDescription(cat.getCategoryDescription());
-		return categoryRepository.save(category);
+		if (categoryRepository.existsById(id)) {
+			CategoryEntity category = categoryRepository.findById(id).get();
+			category.setCategoryName(cat.getCategoryName());
+			category.setCategoryDescription(cat.getCategoryDescription());
+			return categoryRepository.save(category);
+		}
+		return null;
 	}
 
-	// Obrisi kategoriju po ID-u
+	// Obrisi kategoriju po ID-u TESTIRAO
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
 	public CategoryEntity deleteCategory(@PathVariable Integer id) {
-		CategoryEntity temp = categoryRepository.findById(id).get();
-		categoryRepository.deleteById(id);
-		return temp;
+		if (categoryRepository.existsById(id)) {
+			CategoryEntity temp = categoryRepository.findById(id).get();
+			categoryRepository.deleteById(id);
+			return temp;
+		}
+		return null;
 	}
 
-	// Vrati kategoriju po ID-u
+	// Vrati kategoriju po ID-u	TESTIRAO
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
 	public CategoryEntity getCategory(@PathVariable Integer id) {
-		return categoryRepository.findById(id).get();
+		if (categoryRepository.existsById(id)) {
+			return categoryRepository.findById(id).get();
+		}
+		return null;
 	}
 
 }
